requirements.txt
# Flask core
Flask==2.3.3
Werkzeug==3.1.3
itsdangerous==2.1.2
Click==8.1.7

# ORM & Migration
SQLAlchemy==2.0.38
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
pymysql==1.1.0

# Auth & Security
flask-login==0.6.3
flask-hashing==1.1
bcrypt==4.1.2

# Forms & validation
Flask-WTF==1.2.1
email-validator==2.1.1

# File upload
Flask-Uploads==0.2.1
Flask-Dropzone==1.6.0

# REST API support
# Additional required packages based on imports
Flask-RESTful==0.3.10  
Flask-CORS==4.0.0      

# Admin panel
flask-admin==1.6.1

# Dev & debug
flask-debugtoolbar==0.16.0
python-dotenv==1.0.0

# WSGI server
gunicorn==21.2.0
